Apply Spotless formatter fixes
Convert string concatenation to SLF4J placeholders in logging statements
and other minor code style improvements

+++ src/main/java/core/waveform/WaveformRenderer.java
@@ -420,7 +420,7 @@ class WaveformRenderer {
                         if (peak <= 0) {
                             yScale = 0;
                         } else {
-                            yScale = ((key.height() / 2) - 1) / peak;
+                            yScale = (((double) key.height() / 2) - 1) / peak;
                             if (Double.isInfinite(yScale) || Double.isNaN(yScale)) {
                                 yScale = 0;
                             }
